Fix retrieval export matching and split large files
