+++ crates/tandem-server/src/http/skills_memory.rs
@@ -18,4 +18,5 @@ use tandem_types::{
 
 include!("skills_memory_parts/part01.rs");
 include!("skills_memory_parts/part02.rs");
+include!("skills_memory_parts/part04.rs");
 include!("skills_memory_parts/part03.rs");
