+++ crates/tandem-server/src/http/tests/memory.rs
@@ -1,3 +1,4 @@
 include!("memory_parts/part01.rs");
 include!("memory_parts/part02.rs");
 include!("memory_parts/part03.rs");
+include!("memory_parts/part04.rs");
